move errors.rs to execution module
