+++ scylla/src/cluster/cluster.rs
@@ -1,12 +1,12 @@
 use crate::cluster::host_filter::HostFilter;
 use crate::cluster::node::Node;
 use crate::connection::{Connection, PoolConfig, VerifiedKeyspaceName};
+use crate::execution::errors::{BadQuery, NewSessionError, QueryError};
 /// Cluster manages up to date information and connections to database nodes
 use crate::frame::response::event::{Event, StatusChangeEvent};
 use crate::prepared_statement::TokenCalculationError;
 use crate::routing::{Shard, Token};
 use crate::session::TABLET_CHANNEL_SIZE;
-use crate::transport::errors::{BadQuery, NewSessionError, QueryError};
 use crate::transport::locator::ReplicaLocator;
 use crate::transport::partitioner::PartitionerName;
