+++ scylla/src/cluster/metadata.rs
@@ -2,11 +2,11 @@ use crate::cluster::host_filter::HostFilter;
 use crate::cluster::node::resolve_contact_points;
 use crate::connection::{Connection, ConnectionConfig, NodeConnectionPool, PoolConfig, PoolSize};
 use crate::deserialize::DeserializeOwnedRow;
+use crate::execution::errors::{DbError, NewSessionError, QueryError};
 use crate::execution::pager::QueryPager;
 use crate::frame::response::event::Event;
 use crate::routing::Token;
 use crate::statement::query::Query;
-use crate::transport::errors::{DbError, NewSessionError, QueryError};
 use crate::utils::parse::{ParseErrorCause, ParseResult, ParserState};
 
 use futures::future::{self, FutureExt};
@@ -31,7 +31,7 @@ use tracing::{debug, error, trace, warn};
 use uuid::Uuid;
 
 use crate::cluster::node::{InternalKnownNode, NodeAddr, ResolvedContactPoint};
-use crate::transport::errors::{
+use crate::execution::errors::{
     KeyspaceStrategyError, KeyspacesMetadataError, MetadataError, PeersMetadataError,
     ProtocolError, TablesMetadataError, UdtMetadataError, ViewsMetadataError,
 };
