+++ scylla/src/execution/pager.rs
@@ -25,14 +25,15 @@ use tokio::sync::mpsc;
 #[allow(deprecated)]
 use crate::cql_to_rust::{FromRow, FromRowError};
 use crate::deserialize::DeserializeOwnedRow;
+use crate::execution::errors::{QueryError, UserRequestError};
 use crate::execution::execution_profile::ExecutionProfileInner;
 use crate::session::RequestSpan;
-use crate::transport::errors::{QueryError, UserRequestError};
 use crate::transport::query_result::ColumnSpecs;
 
 use crate::cluster::ClusterData;
 use crate::cluster::NodeRef;
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
+use crate::execution::errors::ProtocolError;
 use crate::execution::history::{self, HistoryListener};
 use crate::execution::load_balancing::{self, RoutingInfo};
 use crate::execution::metrics::Metrics;
@@ -42,7 +43,6 @@ use crate::frame::response::{
     result::{ColumnSpec, Row},
 };
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
-use crate::transport::errors::ProtocolError;
 use tracing::{trace, trace_span, warn, Instrument};
 use uuid::Uuid;
 
@@ -81,7 +81,7 @@ mod checked_channel_sender {
     use tokio::sync::mpsc;
     use uuid::Uuid;
 
-    use crate::transport::errors::QueryError;
+    use crate::execution::errors::QueryError;
 
     use super::ReceivedPage;
