+++ scylla/src/execution/retries/downgrading_consistency_rp.rs
@@ -2,7 +2,7 @@ use scylla_cql::Consistency;
 use tracing::debug;
 
 use super::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
-use crate::transport::errors::{DbError, QueryError, WriteType};
+use crate::execution::errors::{DbError, QueryError, WriteType};
 
 /// Downgrading consistency retry policy - retries with lower consistency level if it knows\
 /// that the initial CL is unreachable. Also, it behaves as [DefaultRetryPolicy](crate::execution::retries::DefaultRetryPolicy)
@@ -178,10 +178,10 @@ impl RetrySession for DowngradingConsistencyRetrySession {
 mod tests {
     use bytes::Bytes;
 
-    use crate::test_utils::setup_tracing;
-    use crate::transport::errors::{
+    use crate::execution::errors::{
         BadQuery, BrokenConnectionErrorKind, ConnectionPoolError, ProtocolError,
     };
+    use crate::test_utils::setup_tracing;
 
     use super::*;
