+++ scylla/tests/integration/batch.rs
@@ -1,9 +1,9 @@
 use scylla::batch::Batch;
 use scylla::batch::BatchType;
+use scylla::execution::errors::QueryError;
 use scylla::frame::frame_errors::BatchSerializationError;
 use scylla::frame::frame_errors::CqlRequestSerializationError;
 use scylla::query::Query;
-use scylla::transport::errors::QueryError;
 
 use crate::utils::create_new_session_builder;
 use crate::utils::setup_tracing;
