+++ scylla/tests/integration/execution_profiles.rs
@@ -80,7 +80,7 @@ impl<const NODE: u8> LoadBalancingPolicy for BoundToPredefinedNodePolicy<NODE> {
         _query: &RoutingInfo,
         _latency: std::time::Duration,
         _node: NodeRef<'_>,
-        _error: &scylla::transport::errors::QueryError,
+        _error: &scylla::execution::errors::QueryError,
     ) {
     }
