+++ scylla/tests/integration/history.rs
@@ -3,12 +3,12 @@ use std::sync::Arc;
 
 use chrono::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, Utc};
 use futures::StreamExt;
+use scylla::execution::errors::QueryError;
 use scylla::execution::history::{
     AttemptResult, HistoryCollector, QueryHistoryResult, StructuredHistory, TimePoint,
 };
 use scylla::frame::response::result::Row;
 use scylla::query::Query;
-use scylla::transport::errors::QueryError;
 
 use crate::utils::{create_new_session_builder, setup_tracing, unique_keyspace_name, PerformDDL};
